refactor: Split key exchange publisher/subscriber into separate files.
