+++ src/stream/encryption/KeyExchangeSubscriber.ts
@@ -1,77 +1,23 @@
 import {
-    StreamMessage, GroupKeyRequest, GroupKeyResponse, GroupKeyErrorResponse, EncryptedGroupKey, Errors
+    StreamMessage, GroupKeyRequest, GroupKeyResponse
 } from 'streamr-client-protocol'
-import mem from 'mem'
 import pMemoize from 'p-memoize'
-
 import { uuid, Defer } from '../../utils'
 import Scaffold from '../../utils/Scaffold'
+import mem from 'mem'
 
 import { validateOptions } from '../utils'
-import EncryptionUtil, { GroupKey, GroupKeyish, StreamMessageProcessingError } from './Encryption'
+import EncryptionUtil, { GroupKey } from './Encryption'
 import type { Subscription } from '../../subscribe'
 import { StreamrClient } from '../../StreamrClient'
 import GroupKeyStore from './GroupKeyStore'
-
-const KEY_EXCHANGE_STREAM_PREFIX = 'SYSTEM/keyexchange'
-
-const { ValidationError } = Errors
-
-export function isKeyExchangeStream(id = '') {
-    return id.startsWith(KEY_EXCHANGE_STREAM_PREFIX)
-}
-
-class InvalidGroupKeyRequestError extends ValidationError {
-    code: string
-    constructor(...args: ConstructorParameters<typeof ValidationError>) {
-        super(...args)
-        this.code = 'INVALID_GROUP_KEY_REQUEST'
-        if (Error.captureStackTrace) {
-            Error.captureStackTrace(this, this.constructor)
-        }
-    }
-}
-
-/*
-class InvalidGroupKeyResponseError extends Error {
-    constructor(...args) {
-        super(...args)
-        this.code = 'INVALID_GROUP_KEY_RESPONSE'
-        if (Error.captureStackTrace) {
-            Error.captureStackTrace(this, this.constructor)
-        }
-    }
-}
-
-class InvalidContentTypeError extends Error {
-    constructor(...args) {
-        super(...args)
-        this.code = 'INVALID_MESSAGE_TYPE'
-        if (Error.captureStackTrace) {
-            Error.captureStackTrace(this, this.constructor)
-        }
-    }
-}
-*/
-
-type Address = string
-type GroupKeyId = string
-
-function getKeyExchangeStreamId(address: Address) {
-    if (isKeyExchangeStream(address)) {
-        return address // prevent ever double-handling
-    }
-    return `${KEY_EXCHANGE_STREAM_PREFIX}/${address.toLowerCase()}`
-}
-
-type GroupKeysSerialized = Record<GroupKeyId, GroupKeyish>
-
-function parseGroupKeys(groupKeys: GroupKeysSerialized = {}): Map<GroupKeyId, GroupKey> {
-    return new Map<GroupKeyId, GroupKey>(Object.entries(groupKeys || {}).map(([key, value]) => {
-        if (!value || !key) { return null }
-        return [key, GroupKey.from(value)]
-    }).filter(Boolean) as [])
-}
+import {
+    GroupKeyId,
+    subscribeToKeyExchangeStream,
+    parseGroupKeys,
+    getKeyExchangeStreamId,
+    KeyExhangeOptions,
+} from './KeyExchangeUtils'
 
 type MessageMatch = (content: any, streamMessage: StreamMessage) => boolean
 
@@ -93,186 +39,6 @@ function waitForSubMessage(sub: Subscription, matchFn: MessageMatch) {
         sub.off('error', task.reject)
     }).catch(() => {}) // important: prevent unchained finally cleanup causing unhandled rejection
     return task
-}
-
-async function subscribeToKeyExchangeStream(client: StreamrClient, onKeyExchangeMessage: (msg: any, streamMessage: StreamMessage) => void) {
-    const { options } = client
-    if ((!options.auth!.privateKey && !options.auth!.ethereum) || !options.keyExchange) {
-        return Promise.resolve()
-    }
-
-    await client.session.getSessionToken() // trigger auth errors if any
-    // subscribing to own keyexchange stream
-    const publisherId = await client.getUserId()
-    const streamId = getKeyExchangeStreamId(publisherId)
-    const sub = await client.subscribe(streamId, onKeyExchangeMessage)
-    sub.on('error', () => {}) // errors should not shut down subscription
-    return sub
-}
-
-async function catchKeyExchangeError(client: StreamrClient, streamMessage: StreamMessage, fn: (...args: any[]) => Promise<void>) {
-    try {
-        return await fn()
-    } catch (error) {
-        const subscriberId = streamMessage.getPublisherId()
-        const msg = streamMessage.getParsedContent()
-        const { streamId, requestId, groupKeyIds } = GroupKeyRequest.fromArray(msg)
-        return client.publish(getKeyExchangeStreamId(subscriberId), new GroupKeyErrorResponse({
-            requestId,
-            streamId,
-            errorCode: error.code || 'UNEXPECTED_ERROR',
-            errorMessage: error.message,
-            groupKeyIds
-        }))
-    }
-}
-
-async function PublisherKeyExhangeSubscription(client: StreamrClient, getGroupKeyStore: (streamId: string) => Promise<GroupKeyStore>) {
-    async function onKeyExchangeMessage(_parsedContent: any, streamMessage: StreamMessage) {
-        return catchKeyExchangeError(client, streamMessage, async () => {
-            if (streamMessage.messageType !== StreamMessage.MESSAGE_TYPES.GROUP_KEY_REQUEST) {
-                return Promise.resolve()
-            }
-
-            // No need to check if parsedContent contains the necessary fields because it was already checked during deserialization
-            const { requestId, streamId, rsaPublicKey, groupKeyIds } = GroupKeyRequest.fromArray(streamMessage.getParsedContent())
-
-            const subscriberId = streamMessage.getPublisherId()
-
-            const groupKeyStore = await getGroupKeyStore(streamId)
-            const isSubscriber = await client.isStreamSubscriber(streamId, subscriberId)
-            const encryptedGroupKeys = (!isSubscriber ? [] : await Promise.all(groupKeyIds.map(async (id) => {
-                const groupKey = await groupKeyStore.get(id)
-                if (!groupKey) {
-                    return null // will be filtered out
-                }
-                const key = EncryptionUtil.encryptWithPublicKey(groupKey.data, rsaPublicKey, true)
-                return new EncryptedGroupKey(id, key)
-            }))).filter(Boolean) as EncryptedGroupKey[]
-
-            client.debug('Publisher: Subscriber requested groupKeys: %d. Got: %d. %o', groupKeyIds.length, encryptedGroupKeys.length, {
-                subscriberId,
-                groupKeyIds,
-                responseKeys: encryptedGroupKeys.map(({ groupKeyId }) => groupKeyId),
-            })
-
-            const response = new GroupKeyResponse({
-                streamId,
-                requestId,
-                encryptedGroupKeys,
-            })
-
-            // hack overriding toStreamMessage method to set correct encryption type
-            const toStreamMessage = response.toStreamMessage.bind(response)
-            response.toStreamMessage = (...args) => {
-                const msg = toStreamMessage(...args)
-                msg.encryptionType = StreamMessage.ENCRYPTION_TYPES.RSA
-                return msg
-            }
-
-            return client.publish(getKeyExchangeStreamId(subscriberId), response)
-        })
-    }
-
-    const sub = await subscribeToKeyExchangeStream(client, onKeyExchangeMessage)
-    sub.on('error', (err: Error | StreamMessageProcessingError) => {
-        if (!('streamMessage' in err)) {
-            return // do nothing
-        }
-
-        // wrap error and translate into ErrorResponse.
-        catchKeyExchangeError(client, err.streamMessage, () => { // eslint-disable-line promise/no-promise-in-callback
-            // rethrow so catchKeyExchangeError handles it
-            throw new InvalidGroupKeyRequestError(err.message)
-        }).catch((unexpectedError) => {
-            sub.emit('error', unexpectedError)
-        })
-    })
-
-    return sub
-}
-
-type KeyExhangeOptions = {
-    groupKeys?: Record<string, GroupKeysSerialized>
-}
-
-export function PublisherKeyExhange(client: StreamrClient, { groupKeys = {} }: KeyExhangeOptions = {}) {
-    let enabled = true
-    const getGroupKeyStore = pMemoize(async (streamId) => {
-        const clientId = await client.getAddress()
-        return new GroupKeyStore({
-            clientId,
-            streamId,
-            groupKeys: [...parseGroupKeys(groupKeys[streamId]).entries()]
-        })
-    }, {
-        cacheKey([maybeStreamId]) {
-            const { streamId } = validateOptions(maybeStreamId)
-            return streamId
-        }
-    })
-
-    let sub: Subscription | undefined
-    const next = Scaffold([
-        async () => {
-            sub = await PublisherKeyExhangeSubscription(client, getGroupKeyStore)
-            return async () => {
-                if (!sub) { return }
-                const cancelTask = sub.cancel()
-                sub = undefined
-                await cancelTask
-            }
-        }
-    ], async () => enabled)
-
-    async function rotateGroupKey(streamId: string) {
-        if (!enabled) { return }
-        const groupKeyStore = await getGroupKeyStore(streamId)
-        await groupKeyStore.rotateGroupKey()
-    }
-
-    async function setNextGroupKey(streamId: string, groupKey: GroupKey) {
-        if (!enabled) { return }
-        const groupKeyStore = await getGroupKeyStore(streamId)
-
-        await groupKeyStore.setNextGroupKey(groupKey)
-    }
-
-    async function useGroupKey(streamId: string) {
-        await next()
-        if (!enabled) { return [] }
-        const groupKeyStore = await getGroupKeyStore(streamId)
-        return groupKeyStore.useGroupKey()
-    }
-
-    async function hasAnyGroupKey(streamId: string) {
-        const groupKeyStore = await getGroupKeyStore(streamId)
-        return !groupKeyStore.isEmpty()
-    }
-
-    async function rekey(streamId: string) {
-        if (!enabled) { return }
-        const groupKeyStore = await getGroupKeyStore(streamId)
-        await groupKeyStore.rekey()
-        await next()
-    }
-
-    return {
-        setNextGroupKey,
-        useGroupKey,
-        rekey,
-        rotateGroupKey,
-        hasAnyGroupKey,
-        async start() {
-            enabled = true
-            return next()
-        },
-        async stop() {
-            pMemoize.clear(getGroupKeyStore)
-            enabled = false
-            return next()
-        }
-    }
 }
 
 async function getGroupKeysFromStreamMessage(streamMessage: StreamMessage, encryptionUtil: EncryptionUtil) {
@@ -540,7 +306,7 @@ export function SubscriberKeyExchange(client: StreamrClient, { groupKeys = {} }:
             const streamId = streamMessage.getStreamId()
             const groupKeyStore = await getGroupKeyStore(streamId)
             // newGroupKey has been converted into GroupKey
-            const newGroupKey: unknown = streamMessage.newGroupKey
+            const { newGroupKey } = streamMessage
             await groupKeyStore.add(newGroupKey as GroupKey)
         },
         async stop() {
